Immutable object transactions
As described in #7, repeatedly performing immutable updates can be very slow.
One part of the problem is that object-path-immutable doesn't provide a facility
for performing multiple operations on an object and then returning a single new
object with all affected paths cloned. For example:

```
let o = { foo: { bar: {} } };
o = objectPathImmutable.set(o, 'foo.bar.applesauce', 'chutney');
o = objectPathImmutable.set(o, 'foo.bar.jam', 'marmelade');
```

Both `foo` and `bar` are cloned twice in the above code. What we would like
instead is to perform an immutable transaction. This patch implements an API
that supports the same operations as object-path-immutable, but performed as a
transaction.

This is partial progress toward fixing #7.