Refresh all mount caches after reducing
When a descendant mounted store of a mounted store updated its own state, that
is also a change to the ancestor mounted store's own state. However, as reported
in #10, we did not update the ancestor's cached own state to reflect that
change.

This patch adds a test reproducing the bug and fixes it by updating the cached
state of *all* mounts after reduction is complete. This does unnecessary work,
but assures correctness.

[Fix #10]